ModeInterferenceResult + NotCompatible
