+++ gradle-suppressible-error-prone/src/main/java/com/palantir/gradle/suppressibleerrorprone/modes/common/ModeInterference.java
@@ -16,9 +16,6 @@
 
 package com.palantir.gradle.suppressibleerrorprone.modes.common;
 
-import com.google.common.base.Preconditions;
-import com.palantir.gradle.suppressibleerrorprone.modes.common.ModeInterference.ModeInterferenceResult.Interference;
-import com.palantir.gradle.suppressibleerrorprone.modes.common.ModeInterference.ModeInterferenceResult.NoInterference;
 import java.util.Map;
 import java.util.Set;
 
@@ -32,30 +29,9 @@ public interface ModeInterference {
      * Identify which of the flags interfere with each other.
      *
      * @param modeNames The flags that are enabled in this run
-     * @return Either the flags that interfere with each other, or throw an exception if they are incompatible
+     * @return Either no interference, the flags that interfere with each other, or why they are incompatible
      */
     ModeInterferenceResult interferesWith(Set<ModeName> modeNames);
-
-    sealed interface ModeInterferenceResult permits NoInterference, Interference {
-        static ModeInterferenceResult noInterference() {
-            return NoInterference.INSTANCE;
-        }
-
-        static ModeInterferenceResult interferenceBetween(Set<ModeName> modes) {
-            return new Interference(modes);
-        }
-
-        enum NoInterference implements ModeInterferenceResult {
-            INSTANCE
-        }
-
-        record Interference(Set<ModeName> interferingModes) implements ModeInterferenceResult {
-            public Interference {
-                Preconditions.checkArgument(
-                        interferingModes.size() > 1, "interference must be between at least 2 modes");
-            }
-        }
-    }
 
     /**
      * Modify the modeCommonOptions if the flags interfere with each other.
@@ -65,8 +41,8 @@ public interface ModeInterference {
      */
     default CommonOptions interfere(Map<ModeName, CommonOptions> modeCommonOptions) {
         throw new IllegalStateException(("The interference for this class '%s' has not been implemented. "
-                        + "This is a logic error by the class author as either `interferesWith` should throw or "
-                        + "this method should be implemented.")
+                        + "This is a logic error by the class author as either `interferesWith` should return"
+                        + "not compatible or this method should be implemented.")
                 .formatted(getClass().getCanonicalName()));
     }
 }
